// Kicker_IQ-model-service-isolation-forest-main/src/types/SessionMetrics.ts
type SessionMetrics = {
    distanceM: number;
    highIntensityRunningM: number;
    highIntensityEvents: number;
    sprintDistanceM: number;
    numberOfSprints: number;
    topSpeedKph: number;
    avgSpeedKph: number;
    accelerations: number;
    decelerations: number;
};

export default SessionMetrics